+++ Sources/TuistGenerator/Linter/TargetLinter.swift
@@ -31,6 +31,7 @@ class TargetLinter: TargetLinting {
         issues.append(contentsOf: lintLibraryHasNoResources(target: target))
         issues.append(contentsOf: lintDeploymentTarget(target: target))
         issues.append(contentsOf: settingsLinter.lint(target: target))
+        issues.append(contentsOf: lintDuplicateDependency(target: target))
 
         target.actions.forEach { action in
             issues.append(contentsOf: targetActionLinter.lint(action))
@@ -169,4 +170,14 @@ class TargetLinter: TargetLinting {
 
         return []
     }
+	
+	private func lintDuplicateDependency(target: Target) -> [LintingIssue] {
+		typealias Occurence = Int
+		var seen: [Dependency: Occurence] = [:]
+		target.dependencies.forEach { seen[$0, default: 0] += 1 }
+		let duplicates = seen.enumerated().filter { $0.element.value > 1 }
+		return duplicates.map {
+			.init(reason: "Target has duplicate '\($0.element.key)' dependency specified", severity: .warning)
+		}
+	}
 }